Student Info View
Added a new page to the studet side. For now it is hooked up to the
logout button for the submission view. For it to work: your username
needs to be in the db, need to change the work around to your last and
first name, and need to input a balance to your username