Admin Can Approve Jobs With Insufficient Balance
When an Admin clicks approve on a file where a student has an
insufficient balance, a JOption pane is generated. The admin can then
choose whether or not they would like to allow the student to run a
negative balance. If yes, the student's balance will appear as negative.
If no, no action takes place.